feat(no-throw-literal): add fixes for no-throw-literal rule
This commit adds fix support to the built-in `no-throw-literal` rule,
as well as a couple test cases to verify functionality.

As with the existing implementation, literal (strings, numbers, etc) and
`undefined` values are supported, generating fixes that wrap the literal in a
new `Error` object, or just return an empty `Error` object, respectively.

In addition to the changes above, the `no-throw-literal` rule now handles
template literals. The previous implementation would simply ignore them,
which didn't make much sense to me.

Tagged template literals are still ignored, as they could be used to create a
custom error object, or any number of other things, and may not necessarily be
a literal value when they are resolved at runtime.

The `void` operator is also now supported, and is handled in the same manner
as `undefined` values. This is a breaking change, as the previous implementation
would just ignore them entirely.

- [x] template literals are wrapped in `new Error()`
 - `` throw `foo` `` -> ``throw new Error(`foo`)``
- [x] tagged template literals are ignored
 - `` throw foo`bar` `` -> `` throw foo`bar` ``
- [x] quote style is preserved around literal values (if present)
 - `throw "foo"` -> `throw new Error("foo")`
 - `throw 'foo'` -> `throw new Error('foo')`
 - `` throw `foo` `` -> ``throw new Error(`foo`)``
- [x] non-string literals are converted to `new Error("<literal>")`
 - `throw 42` -> `throw new Error("42")`
 - `throw true` -> `throw new Error("true")`
 - `throw false` -> `throw new Error("false")`
 - `throw null` -> `throw new Error("null")`
- [x] `undefined` values are replaced with `new Error()`
- [x] `void <expression>` values are replaced with `new Error()`

Thanks!
Nick